further split prepare into 3 other files
